refactor(client): decompose CombatPanel into focused sub-components
Extract 1,626-line monolith into 8 focused files under combat/:
- StyleIcons.tsx: SVG icon components for attack styles
- CombatStyleSelector.tsx: style banner grid with draggable shields
- CombatBonusesDisplay.tsx: HP bar, combat level, stat rows
- AutoRetaliateToggle.tsx: OSRS-style on/off toggle
- SpecialAttackBar.tsx: special attack energy bar
- types.ts: shared interfaces for combat sub-components
- typeGuards.ts: runtime event validation

CombatPanel.tsx reduced to 448 lines as orchestrator managing state
and world event subscriptions. Zero behavior changes.
