+++ packages/client/src/game/panels/combat/index.ts
@@ -1,0 +1,35 @@
+/**
+ * Combat Panel Sub-Components
+ *
+ * Barrel export for all combat panel components and types.
+ */
+
+export {
+  StyleIcon,
+  StatIcon,
+  BannerStyleIcon,
+  SHIELD_OUTER,
+  SHIELD_INNER,
+  XP_SHORT_LABELS,
+} from "./StyleIcons";
+export { CombatStyleSelector } from "./CombatStyleSelector";
+export { CombatBonusesDisplay } from "./CombatBonusesDisplay";
+export { SpecialAttackBar } from "./SpecialAttackBar";
+export { AutoRetaliateToggle } from "./AutoRetaliateToggle";
+export {
+  isStyleUpdateEvent,
+  isTargetChangedEvent,
+  isTargetHealthEvent,
+  isAutoRetaliateEvent,
+} from "./typeGuards";
+export type {
+  CombatStyleInfo,
+  CombatStyleSelectorProps,
+  CombatBonusesDisplayProps,
+  SpecialAttackBarProps,
+  AutoRetaliateToggleProps,
+  StyleUpdateEvent,
+  TargetChangedEvent,
+  TargetHealthEvent,
+  AutoRetaliateEvent,
+} from "./types";
