+++ packages/client/src/game/panels/combat/typeGuards.ts
@@ -1,0 +1,49 @@
+/**
+ * Combat Panel Type Guards
+ *
+ * Runtime validation functions for combat event data.
+ */
+
+import type {
+  StyleUpdateEvent,
+  TargetChangedEvent,
+  TargetHealthEvent,
+  AutoRetaliateEvent,
+} from "./types";
+
+export function isStyleUpdateEvent(data: unknown): data is StyleUpdateEvent {
+  if (typeof data !== "object" || data === null) return false;
+  const d = data as Record<string, unknown>;
+  return (
+    typeof d.playerId === "string" &&
+    typeof d.currentStyle === "object" &&
+    d.currentStyle !== null &&
+    typeof (d.currentStyle as Record<string, unknown>).id === "string"
+  );
+}
+
+export function isTargetChangedEvent(
+  data: unknown,
+): data is TargetChangedEvent {
+  if (typeof data !== "object" || data === null) return false;
+  const d = data as Record<string, unknown>;
+  return d.targetId === null || typeof d.targetId === "string";
+}
+
+export function isTargetHealthEvent(data: unknown): data is TargetHealthEvent {
+  if (typeof data !== "object" || data === null) return false;
+  const d = data as Record<string, unknown>;
+  return (
+    typeof d.targetId === "string" &&
+    typeof d.health === "object" &&
+    d.health !== null
+  );
+}
+
+export function isAutoRetaliateEvent(
+  data: unknown,
+): data is AutoRetaliateEvent {
+  if (typeof data !== "object" || data === null) return false;
+  const d = data as Record<string, unknown>;
+  return typeof d.playerId === "string" && typeof d.enabled === "boolean";
+}
